Add the repoquerywrapper tool. (#6390)
* Add the repoquerywrapper tool.

The repoquerywrapper tools creates a chroot, installs dnf packages, and
runs specific repoquery commands within that environment.

This tools allows us to run repoquery on systems that may not have repoquery
installed/available.

It also allows us to run queries that are not directly supported by repoquery -
for example, in this check-in, it allows us to check which packages in a list
are available on the target repros.

Such queries are very specific - and can be added incrementally to the tool as
the need arises.

* Fix formatting (make go-tidy-all).

* Fix typo in help message.
